Gadfly

Gadfly

Compose.draw

Compose.gridstack

Compose.hstack

Compose.vstack

Gadfly.current_theme

Gadfly.get_theme

Gadfly.get_theme

Gadfly.get_theme

Gadfly.lab_gradient

Gadfly.lab_rainbow

Gadfly.layer

Gadfly.layer

Gadfly.layer

Gadfly.layer

Gadfly.layer

Gadfly.lchabmix

Gadfly.luv_rainbow

Gadfly.plot

Gadfly.plot

Gadfly.plot

Gadfly.plot

Gadfly.plot

Gadfly.plot

Gadfly.pop_theme

Gadfly.push_theme

Gadfly.push_theme

Gadfly.render

Gadfly.set_default_plot_format

Gadfly.set_default_plot_size

Gadfly.spy

Gadfly.style

Gadfly.title

Gadfly.weighted_color_mean

Gadfly.with_theme

Compose.draw — Method.

draw(backend::Compose.Backend, p::Plot)

A convenience version of Compose.draw without having to call render.

source

Compose.gridstack — Method.

gridstack(ps::Matrix{Union{Plot,Context}})

Arrange plots into a rectangular array. Use context() as a placeholder for an empty panel. Heterogeneous matrices must be typed. See also hstack and vstack.

Examples

p1 = plot(x=[1,2], y=[3,4], Geom.line);
p2 = Compose.context();
gridstack([p1 p1; p1 p1])
gridstack(Union{Plot,Compose.Context}[p1 p2; p2 p1])

source

Compose.hstack — Method.

hstack(ps::Union{Plot,Context}...)
hstack(ps::Vector)

Arrange plots into a horizontal row. Use context() as a placeholder for an empty panel. Heterogeneous vectors must be typed. See also vstack, gridstack, and Geom.subplot_grid.

Examples

p1 = plot(x=[1,2], y=[3,4], Geom.line);
p2 = Compose.context();
hstack(p1, p2)
hstack(Union{Plot,Compose.Context}[p1, p2])

source

Compose.vstack — Method.

vstack(ps::Union{Plot,Context}...)
vstack(ps::Vector)

Arrange plots into a vertical column. Use context() as a placeholder for an empty panel. Heterogeneous vectors must be typed. See also hstack, gridstack, and Geom.subplot_grid.

Examples

p1 = plot(x=[1,2], y=[3,4], Geom.line);
p2 = Compose.context();
vstack(p1, p2)
vstack(Union{Plot,Compose.Context}[p1, p2])

source

Gadfly.layer — Method.

layer(f::Function, xmin::Number, xmax::Number, ymin::Number, ymax::Number,
      elements::ElementOrFunction...; mapping...) -> [Layers]

Create a layer of the contours of the 2D function or expression in f. See Stat.func and Geom.contour.

source

Gadfly.layer — Method.

layer(f::Function, lower::Number, upper::Number,
      elements::ElementOrFunction...) -> [Layers]

Create a layer from the function or expression f, which takes a single argument or operates on a single variable, respectively, between the lower and upper bounds. See Stat.func and Geom.line.

source

Gadfly.layer — Method.

layer(data_source::Union{AbstractDataFrame, Void}),
      elements::ElementOrFunction...; mapping...) -> [Layers]

Create a layer element based on the data in data_source, to later input into plot. elements can be Statistics, Geometries, and/or Themes (but not Scales, Coordinates, or Guides). mapping are aesthetics.

Examples

ls=[]
append!(ls, layer(y=[1,2,3], Geom.line))
append!(ls, layer(y=[3,2,1], Geom.point))
plot(ls..., Guide.title("layer example"))

source

Gadfly.layer — Method.

layer(elements::ElementOrFunction...; mapping...) =
      layer(nothing, elements...; mapping...) -> [Layers]

source

Gadfly.layer — Method.

layer(fs::Vector{T}, lower::Number, upper::Number,
      elements::ElementOrFunction...) where T <: Base.Callable -> [Layers]

Create a layer from a list of functions or expressions in fs.

source

Gadfly.plot — Method.

plot(f::Function, xmin::Number, xmax::Number, ymin::Number, ymax::Number,
     elements::ElementOrFunction...; mapping...)

Plot the contours of the 2D function or expression in f. See Stat.func and Geom.contour.

source

Gadfly.plot — Method.

plot(f::Function, lower::Number, upper::Number, elements::ElementOrFunction...;
     mapping...)

Plot the function or expression f, which takes a single argument or operates on a single variable, respectively, between the lower and upper bounds. See Stat.func and Geom.line.

source

Gadfly.plot — Method.

plot(data_source::Union{Void, AbstractMatrix, AbstractDataFrame},
     mapping::Dict, elements::ElementOrFunctionOrLayers...) -> Plot

The old fashioned (pre-named arguments) version of plot. This version takes an explicit mapping dictionary, mapping aesthetics symbols to expressions or columns in the data frame.

source

Gadfly.plot — Method.

plot(data_source::Union{AbstractMatrix, AbstractDataFrame},
     elements::ElementOrFunctionOrLayers...; mapping...) -> Plot

Create a new plot by specifying a data_source, one or more plot elements (Scales, Statistics, Coordinates, Geometries, Guides, Themes, and/or Layers), and a mapping of aesthetics to columns or expressions of the data.

Examples

my_data = DataFrame(time=1917:2018, price=1.02.^(0:101))
plot(my_data, Geom.line, x=:time, y=:price)

# or equivalently:
plot(Geom.line, x=collect(1917:2018), y=1.02.^(0:101))

source

Gadfly.plot — Method.

plot(elements::ElementOrFunctionOrLayers...; mapping...) -> Plot

source

Gadfly.plot — Method.

plot(fs::Vector{T}, lower::Number, upper::Number, elements::ElementOrFunction...;
     mapping...) where T <: Base.Callable

source

Gadfly.render — Method.

render(plot::Plot) -> Context

Render plot to a Compose context.

source

Gadfly.set_default_plot_format — Method.

set_default_plot_format(fmt::Symbol)

Sets the default plot format.

source

Gadfly.set_default_plot_size — Method.

set_default_plot_size(width::Compose.MeasureOrNumber,
                      height::Compose.MeasureOrNumber)

Sets preferred canvas size when rendering a plot without an explicit call to draw. Units can be inch, cm, mm, pt, or px.

source

Gadfly.spy — Method.

spy(M::AbstractMatrix, elements::ElementOrFunction...; mapping...) -> Plot

Plots a heatmap of M, with M[1,1] in the upper left. NaN values are left blank, and an error is thrown if all elements of M are NaN. See Geom.rectbin and Coord.cartesian(fixed=true)...).

source

Gadfly.style — Method.

style(; kwargs...) -> Theme

Return a new Theme that is a copy of the current theme as modifed by the attributes in kwargs. See Themes for available fields.

Examples

style(background_color="gray")

source

Gadfly.title — Method.

title(ctx::Context, str::String, props::Property...) -> Context

Add a title string to a group of plots, typically created with vstack, hstack, or gridstack.

Examples

p1 = plot(x=[1,2], y=[3,4], Geom.line);
p2 = plot(x=[1,2], y=[4,3], Geom.line);
title(hstack(p1,p2), "my latest data", Compose.fontsize(18pt), fill(colorant"red"))

source

Base.Multimedia.display — Method.

display(p::Plot)

Render p to a multimedia display, typically an internet browser. This function is handy when rendering by plot has been suppressed with either trailing semi-colon or by calling it within a function.

source

Gadfly.current_theme — Method.

current_theme()

Get the Theme on top of the theme stack.

source

Gadfly.get_theme — Method.

get_theme(::Val{:dark})

A light foreground on a dark background.

source

Gadfly.get_theme — Method.

get_theme(::Val{:default})

A dark foreground on a light background.

source

Gadfly.get_theme — Method.

get_theme()

Register a theme by name by adding methods to get_theme.

Examples

get_theme(::Val{:mytheme}) = Theme(...)
push_theme(:mytheme)

source

Gadfly.lab_gradient — Method.

function lab_gradient(cs::Color...)

source

Gadfly.lab_rainbow — Method.

lab_rainbow(l, c, h0, n)

Generate n colors in the LCHab colorspace by using a fixed luminance l and chroma c, and varying the hue, starting at h0.

source

Gadfly.lchabmix — Method.

function lchabmix(c0_, c1_, r, power)

source

Gadfly.luv_rainbow — Method.

luv_rainbow(l, c, h0, n)

Generate n colors in the LCHuv colorspace by using a fixed luminance l and chroma c, and varying the hue, starting at h0.

source

Gadfly.pop_theme — Method.

pop_theme() -> Theme

Return to using the previous theme by removing the top item on the theme stack. See also pop_theme and with_theme.

source

Gadfly.push_theme — Method.

push_theme(t::Theme)

Set the current theme by placing t onto the top of the theme stack. See also pop_theme and with_theme.

source

Gadfly.push_theme — Method.

push_theme(t::Symbol)

Push a Theme by its name. Available options are :default and :dark. See also get_theme.

source

Gadfly.weighted_color_mean — Method.

function weighted_color_mean(cs::AbstractArray{Lab{T},1},
                             ws::AbstractArray{S,1}) where {S <: Number,T}

Return the mean of Lab colors cs as weighted by ws.

source

Gadfly.with_theme — Method.

with_theme(f, theme)

Call function f with theme as the current Theme. theme can be a Theme object or a symbol.

Examples

with_theme(style(background_color=colorant"#888888"))) do
    plot(x=rand(10), y=rand(10))
end

source