Guides

Guides

Very similar to Geometries are guides, which draw graphics supporting the actual visualization, such as axis ticks, axis labels, and color keys. The major distinction is that geometries always draw within the rectangular plot frame, while guides have some special layout considerations.

Gadfly.Guide.annotation

Gadfly.Guide.colorkey

Gadfly.Guide.manual_color_key

Gadfly.Guide.shapekey

Gadfly.Guide.title

Gadfly.Guide.xlabel

Gadfly.Guide.xrug

Gadfly.Guide.xticks

Gadfly.Guide.ylabel

Gadfly.Guide.yrug

Gadfly.Guide.yticks

Gadfly.Guide.annotation — Type.

Guide.annotation(ctx::Compose.Context)

Overlay a plot with an arbitrary Compose graphic. The context will inherit the plot's coordinate system, unless overridden with a custom unit box.

source

Gadfly.Guide.colorkey — Type.

Guide.colorkey[(; title=nothing, labels=nothing, pos=nothing)]
Guide.colorkey(title, labels, pos)

Enable control of the auto-generated colorkey. Set the colorkey title for any plot, and the item labels for plots with a discrete color scale. pos overrides Theme(key_position=) and can be in either relative (e.g. [0.7w, 0.2h] is the lower right quadrant), absolute (e.g. [0mm, 0mm]), or plot scale (e.g. [0,0]) coordinates.

source

Gadfly.Guide.manual_color_key — Type.

Guide.manual_color_key(title, labels, colors)

Manually define a color key with the legend title and item labels and colors.

source

Gadfly.Guide.shapekey — Type.

Guide.shapekey[(; title="Shape", labels=[""], pos=Float64[])]
Guide.shapekey(title, labels, pos)

Enable control of the auto-generated shapekey. Set the key title and the item labels. pos overrides Theme(key_position=) and can be in either relative (e.g. [0.7w, 0.2h] is the lower right quadrant), absolute (e.g. [0mm, 0mm]), or plot scale (e.g. [0,0]) coordinates.

source

Gadfly.Guide.title — Type.

Geom.title(title)

Set the plot title.

source

Gadfly.Guide.xlabel — Type.

Guide.xlabel(label, orientation=:auto)

Sets the x-axis label for the plot. label is either a String or nothing. orientation can also be :horizontal or :vertical.

source

Gadfly.Guide.xrug — Type.

Guide.xrug

Draw a short vertical lines along the x-axis of a plot at the positions in the x aesthetic.

source

Gadfly.Guide.xticks — Type.

Guide.xticks[(; label=true, ticks=:auto, orientation=:auto)]
Guide.xticks(label, ticks, orientation)

Formats the tick marks and labels for the x-axis. label toggles the label visibility. ticks can also be an array of locations, or nothing. orientation can also be :horizontal or :vertical.

source

Gadfly.Guide.ylabel — Type.

Guide.ylabel(label, orientation=:auto)

Sets the y-axis label for the plot. label is either a String or nothing. orientation can also be :horizontal or :vertical.

source

Gadfly.Guide.yrug — Type.

Guide.yrug

Draw short horizontal lines along the y-axis of a plot at the positions in the 'y' aesthetic.

source

Gadfly.Guide.yticks — Type.

Guide.yticks[(; label=true, ticks=:auto, orientation=:horizontal)]
Guide.yticks(ticks, label, orientation)

Formats the tick marks and labels for the y-axis. label toggles the label visibility. ticks can also be an array of locations, or nothing. orientation can also be :auto or :vertical.

source