Shapes

Shapes, when combined with Geom.point, specify the appearance of markers. In addition to those below, circle is also imported from Compose.jl.

Gadfly.Shape.cross

Gadfly.Shape.diamond

Gadfly.Shape.dtriangle

Gadfly.Shape.hexagon

Gadfly.Shape.hline

Gadfly.Shape.ltriangle

Gadfly.Shape.octagon

Gadfly.Shape.rtriangle

Gadfly.Shape.square

Gadfly.Shape.star1

Gadfly.Shape.star2

Gadfly.Shape.utriangle

Gadfly.Shape.vline

Gadfly.Shape.xcross

Gadfly.Shape.cross — Method

cross(xs, ys, rs)

Draw crosses at the coordinates specified in xs and ys of size rs

Gadfly.Shape.diamond — Method

diamond(xs, ys, rs)

Draw diamonds at the coordinates specified in xs and ys of size rs

Gadfly.Shape.dtriangle — Method

dtriangle(xs, ys, rs)

Draw downward-pointing triangles at the coordinates specified in xs and ys of size rs

Gadfly.Shape.hexagon — Method

hexagon(xs, ys, rs)

Draw hexagons at the coordinates specified in xs and ys of size rs

Gadfly.Shape.hline — Function

hline(xs, ys, rs, scalar=1)

Draw horizontal lines at the coordinates specified in xs and ys of size rs

Gadfly.Shape.ltriangle — Function

ltriangle(xs, ys, rs, scalar=1)

Draw left-pointing triangles at the coordinates specified in xs and ys of size rs

Gadfly.Shape.octagon — Method

octagon(xs, ys, rs)

Draw octagons at the coordinates specified in xs and ys of size rs

Gadfly.Shape.rtriangle — Method

rtriangle(xs, ys, rs)

Draw right-pointing triangles at the coordinates specified in xs and ys of size rs

Gadfly.Shape.square — Method

square(xs, ys, rs)

Draw squares at the coordinates specified in xs and ys of size rs

Gadfly.Shape.star1 — Function

star1(xs, ys, rs, scalar=1)

Draw five-pointed stars at the coordinates specified in xs and ys of size rs

Gadfly.Shape.star2 — Function

star2(xs, ys, rs, scalar=1)

Draw four-pointed stars at the coordinates specified in xs and ys of size rs

Gadfly.Shape.utriangle — Function

utriangle(xs, ys, rs, scalar=1)

Draw upward-pointing triangles at the coordinates specified in xs and ys of size rs

Gadfly.Shape.vline — Function

vline(xs, ys, rs, scalar=1)

Draw vertical lines at the coordinates specified in xs and ys of size rs

Gadfly.Shape.xcross — Method

xcross(xs, ys, rs)

Draw rotated crosses at the coordinates specified in xs and ys of size rs