Coordinates

Coordinate systems are mappings between a coordinate space and the 2D rendered output. Currently there is only 2D Cartesian, but this would be the mechanism to implement polar, barycentric, etc. and even projections of their 3D counterparts.

Gadfly.Coord.cartesian

Gadfly.Coord.cartesian — Type

Coord.cartesian(; xmin=nothing, xmax=nothing, ymin=nothing, ymax=nothing,
                xflip=false, yflip=false,
                aspect_ratio=nothing, fixed=false,
                raster=false)

xmin, xmax, ymin, and ymax specify hard minimum and maximum values on the x and y axes, and override the soft limits in Scale.x_continuous and Scale.y_continuous. if xflip or yflip are true the respective axis is flipped. aspect_ratio fulfills its namesake if not nothing, unless overridden by a fixed value of true, in which case the aspect ratio follows the units of the plot (e.g. if the y-axis is 5 units high and the x-axis in 10 units across, the plot will be drawn at an aspect ratio of 2).